Fix review findings: extract shared utils, remove unused imports
- Extract countProseWords() and getEntityTypeFromPath() into
  crux/lib/page-analysis.ts to eliminate duplication across
  citation-density, balance-flags, and validate-hallucination-risk
- Remove unused imports (existsSync, readdirSync, parseYaml) from
  validate-hallucination-risk.ts
